Extract route handlers from main.py into focused API modules
main.py was 874 lines and growing toward god-file territory. Split route
handlers into four focused modules under gateway/api/:

- api/query.py (198 lines) — governed SQL execution with cost estimation,
  PII redaction, and caching
- api/sandboxes.py (111 lines) — Firecracker sandbox CRUD and code execution
- api/audit.py (98 lines) — audit trail query and CSV/JSON export
- api/governance.py (185 lines) — budget, cache, annotations, PII detection

main.py is now 351 lines containing only app setup, middleware, health,
settings, connections, and metrics. Each module has a single responsibility.
