+++ signalpilot/gateway/gateway/api/sandboxes.py
@@ -1,0 +1,111 @@
+"""Sandbox routes — create, manage, and execute code in Firecracker microVMs."""
+
+from __future__ import annotations
+
+import time
+import uuid
+
+from fastapi import APIRouter, HTTPException
+
+from ..models import AuditEntry, ExecuteRequest, SandboxCreate
+from ..store import (
+    append_audit,
+    delete_sandbox,
+    get_sandbox,
+    list_sandboxes,
+    load_settings,
+    upsert_sandbox,
+)
+
+router = APIRouter()
+
+
+def _get_sandbox_client():
+    from ..main import _get_sandbox_client as _get_client
+    return _get_client()
+
+
+@router.get("/api/sandboxes")
+async def get_sandboxes():
+    return list_sandboxes()
+
+
+@router.post("/api/sandboxes", status_code=201)
+async def create_sandbox(req: SandboxCreate):
+    session_token = str(uuid.uuid4())
+
+    client = _get_sandbox_client()
+    sandbox = await client.create_sandbox(
+        session_token=session_token,
+        connection_name=req.connection_name,
+        label=req.label,
+        budget_usd=req.budget_usd,
+        row_limit=req.row_limit,
+    )
+    upsert_sandbox(sandbox)
+
+    await append_audit(AuditEntry(
+        id=str(uuid.uuid4()),
+        timestamp=time.time(),
+        event_type="connect",
+        connection_name=req.connection_name,
+        sandbox_id=sandbox.id,
+        metadata={"label": req.label},
+    ))
+
+    return sandbox
+
+
+@router.get("/api/sandboxes/{sandbox_id}")
+async def get_sandbox_detail(sandbox_id: str):
+    sandbox = get_sandbox(sandbox_id)
+    if not sandbox:
+        raise HTTPException(status_code=404, detail="Sandbox not found")
+    return sandbox
+
+
+@router.delete("/api/sandboxes/{sandbox_id}", status_code=204)
+async def kill_sandbox(sandbox_id: str):
+    sandbox = get_sandbox(sandbox_id)
+    if not sandbox:
+        raise HTTPException(status_code=404, detail="Sandbox not found")
+
+    if sandbox.vm_id:
+        client = _get_sandbox_client()
+        await client.kill(sandbox.vm_id)
+
+    delete_sandbox(sandbox_id)
+
+
+@router.post("/api/sandboxes/{sandbox_id}/execute")
+async def execute_in_sandbox(sandbox_id: str, req: ExecuteRequest):
+    sandbox = get_sandbox(sandbox_id)
+    if not sandbox:
+        raise HTTPException(status_code=404, detail="Sandbox not found")
+
+    if sandbox.status == "stopped":
+        raise HTTPException(status_code=409, detail="Sandbox has been stopped")
+
+    session_token = str(uuid.uuid4())
+
+    client = _get_sandbox_client()
+    result = await client.execute(
+        sandbox=sandbox,
+        code=req.code,
+        session_token=session_token,
+        timeout=req.timeout,
+    )
+
+    # Update sandbox state
+    upsert_sandbox(sandbox)
+
+    await append_audit(AuditEntry(
+        id=str(uuid.uuid4()),
+        timestamp=time.time(),
+        event_type="execute",
+        connection_name=sandbox.connection_name,
+        sandbox_id=sandbox_id,
+        metadata={"code_preview": req.code[:200], "success": result.success},
+    ))
+
+    return result
