+++ signalpilot/gateway/gateway/main.py
@@ -27,52 +27,39 @@ from __future__ import annotations
 
 import asyncio
 import json
+import os
+import re
 import time
-import uuid
 from contextlib import asynccontextmanager
 
 from fastapi import FastAPI, HTTPException, Query
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import StreamingResponse
 
-from .engine import inject_limit, validate_sql
 from .middleware import APIKeyAuthMiddleware, RateLimitMiddleware, SecurityHeadersMiddleware
-from .models import (
-    AuditEntry,
-    ConnectionCreate,
-    ExecuteRequest,
-    GatewaySettings,
-    SandboxCreate,
-)
+from .models import ConnectionCreate, GatewaySettings
 from .connectors.pool_manager import pool_manager
 from .connectors.health_monitor import health_monitor
 from .connectors.schema_cache import schema_cache
 from .sandbox_client import SandboxClient
 from .store import (
-    append_audit,
     create_connection,
     delete_connection,
-    delete_sandbox,
     get_connection,
     get_connection_string,
-    get_sandbox,
     list_connections,
     list_sandboxes,
     load_settings,
-    read_audit,
     save_settings,
-    upsert_sandbox,
 )
 
 # ─── Error Sanitization (HIGH-06) ────────────────────────────────────────────
 
-import re as _re
-
 _SENSITIVE_PATTERNS = [
-    _re.compile(r"postgresql://[^\s]+", _re.IGNORECASE),
-    _re.compile(r"mysql://[^\s]+", _re.IGNORECASE),
-    _re.compile(r"password[=:]\s*\S+", _re.IGNORECASE),
-    _re.compile(r"host=\S+", _re.IGNORECASE),
+    re.compile(r"postgresql://[^\s]+", re.IGNORECASE),
+    re.compile(r"mysql://[^\s]+", re.IGNORECASE),
+    re.compile(r"password[=:]\s*\S+", re.IGNORECASE),
+    re.compile(r"host=\S+", re.IGNORECASE),
 ]
 
 
@@ -81,7 +68,6 @@ def _sanitize_db_error(error: str) -> str:
     sanitized = error
     for pattern in _SENSITIVE_PATTERNS:
         sanitized = pattern.sub("[REDACTED]", sanitized)
-    # Truncate to prevent information dump
     if len(sanitized) > 500:
         sanitized = sanitized[:500] + "..."
     return sanitized
@@ -114,7 +100,6 @@ def _reset_sandbox_client():
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
-    # Start background pool cleanup task
     async def _pool_cleanup_loop():
         while True:
             await asyncio.sleep(60)
@@ -144,9 +129,7 @@ _ALLOWED_ORIGINS = [
     "http://127.0.0.1:3200",
     "http://127.0.0.1:3000",
 ]
-# Allow override via env var for production deployments
-import os as _os
-_extra_origins = _os.getenv("SP_ALLOWED_ORIGINS", "")
+_extra_origins = os.getenv("SP_ALLOWED_ORIGINS", "")
 if _extra_origins:
     _ALLOWED_ORIGINS.extend(o.strip() for o in _extra_origins.split(",") if o.strip())
 
@@ -162,6 +145,18 @@ app.add_middleware(
 app.add_middleware(SecurityHeadersMiddleware)
 app.add_middleware(RateLimitMiddleware, general_rpm=120, expensive_rpm=30)
 app.add_middleware(APIKeyAuthMiddleware)
+
+# ─── Include route modules ──────────────────────────────────────────────────
+
+from .api.query import router as query_router
+from .api.sandboxes import router as sandboxes_router
+from .api.audit import router as audit_router
+from .api.governance import router as governance_router
+
+app.include_router(query_router)
+app.include_router(sandboxes_router)
+app.include_router(audit_router)
+app.include_router(governance_router)
 
 
 # ─── Health ──────────────────────────────────────────────────────────────────
@@ -197,7 +192,7 @@ async def get_settings():
 @app.put("/api/settings")
 async def update_settings(settings: GatewaySettings):
     save_settings(settings)
-    _reset_sandbox_client()  # Reconnect with new URL
+    _reset_sandbox_client()
     return settings
 
 
@@ -279,7 +274,6 @@ async def get_connection_schema(name: str):
     if not conn_str:
         raise HTTPException(status_code=400, detail="No credentials stored for this connection")
 
-    # Check schema cache first (Feature #18)
     cached = schema_cache.get(name)
     if cached is not None:
         return {
@@ -299,7 +293,6 @@ async def get_connection_schema(name: str):
     except Exception as e:
         raise HTTPException(status_code=500, detail=_sanitize_db_error(str(e)))
 
-    # Store in cache
     schema_cache.put(name, schema)
     return {
         "connection_name": name,
@@ -310,525 +303,10 @@ async def get_connection_schema(name: str):
     }
 
 
-# ─── Sandboxes ───────────────────────────────────────────────────────────────
+# ─── Metrics SSE ─────────────────────────────────────────────────────────────
 
-@app.get("/api/sandboxes")
-async def get_sandboxes():
-    return list_sandboxes()
-
-
-@app.post("/api/sandboxes", status_code=201)
-async def create_sandbox(req: SandboxCreate):
-    session_token = str(uuid.uuid4())
-    settings = load_settings()
-
-    client = _get_sandbox_client()
-    sandbox = await client.create_sandbox(
-        session_token=session_token,
-        connection_name=req.connection_name,
-        label=req.label,
-        budget_usd=req.budget_usd,
-        row_limit=req.row_limit,
-    )
-    upsert_sandbox(sandbox)
-
-    await append_audit(AuditEntry(
-        id=str(uuid.uuid4()),
-        timestamp=time.time(),
-        event_type="connect",
-        connection_name=req.connection_name,
-        sandbox_id=sandbox.id,
-        metadata={"label": req.label},
-    ))
-
-    return sandbox
-
-
-@app.get("/api/sandboxes/{sandbox_id}")
-async def get_sandbox_detail(sandbox_id: str):
-    sandbox = get_sandbox(sandbox_id)
-    if not sandbox:
-        raise HTTPException(status_code=404, detail="Sandbox not found")
-    return sandbox
-
-
-@app.delete("/api/sandboxes/{sandbox_id}", status_code=204)
-async def kill_sandbox(sandbox_id: str):
-    sandbox = get_sandbox(sandbox_id)
-    if not sandbox:
-        raise HTTPException(status_code=404, detail="Sandbox not found")
-
-    if sandbox.vm_id:
-        client = _get_sandbox_client()
-        await client.kill(sandbox.vm_id)
-
-    delete_sandbox(sandbox_id)
-
-
-@app.post("/api/sandboxes/{sandbox_id}/execute")
-async def execute_in_sandbox(sandbox_id: str, req: ExecuteRequest):
-    sandbox = get_sandbox(sandbox_id)
-    if not sandbox:
-        raise HTTPException(status_code=404, detail="Sandbox not found")
-
-    if sandbox.status == "stopped":
-        raise HTTPException(status_code=409, detail="Sandbox has been stopped")
-
-    settings = load_settings()
-    session_token = str(uuid.uuid4())  # In production, this is tied to the session
-
-    client = _get_sandbox_client()
-    result = await client.execute(
-        sandbox=sandbox,
-        code=req.code,
-        session_token=session_token,
-        timeout=req.timeout,
-    )
-
-    # Update sandbox state
-    upsert_sandbox(sandbox)
-
-    await append_audit(AuditEntry(
-        id=str(uuid.uuid4()),
-        timestamp=time.time(),
-        event_type="execute",
-        connection_name=sandbox.connection_name,
-        sandbox_id=sandbox_id,
-        metadata={"code_preview": req.code[:200], "success": result.success},
-    ))
-
-    return result
-
-
-# ─── Direct SQL Query ─────────────────────────────────────────────────────────
-
-class QueryRequest(ConnectionCreate.__class__):
-    pass
-
-
-from pydantic import BaseModel, Field
-
-
-class DirectQueryRequest(BaseModel):
-    connection_name: str = Field(..., min_length=1, max_length=64, pattern=r"^[a-zA-Z0-9_-]+$")
-    sql: str = Field(..., min_length=1, max_length=100_000)
-    row_limit: int = Field(default=10_000, ge=1, le=100_000)
-    timeout_seconds: int | None = Field(default=None, ge=1, le=300)
-
-
-@app.post("/api/query")
-async def query_database(req: DirectQueryRequest):
-    info = get_connection(req.connection_name)
-    if not info:
-        raise HTTPException(status_code=404, detail=f"Connection '{req.connection_name}' not found")
-
-    settings = load_settings()
-    timeout = req.timeout_seconds or settings.default_timeout_seconds
-
-    # Load annotations for blocked tables check (Feature #19)
-    annotations = load_annotations(req.connection_name)
-    blocked_tables = list(annotations.blocked_tables)
-
-    # Merge with settings-level blocked tables
-    if settings.blocked_tables:
-        blocked_tables.extend(t for t in settings.blocked_tables if t not in blocked_tables)
-
-    # Validate SQL (with blocked tables from annotations + settings)
-    validation = validate_sql(req.sql, blocked_tables=blocked_tables or None)
-    if not validation.ok:
-        await append_audit(AuditEntry(
-            id=str(uuid.uuid4()),
-            timestamp=time.time(),
-            event_type="block",
-            connection_name=req.connection_name,
-            sql=req.sql,
-            blocked=True,
-            block_reason=validation.blocked_reason,
-        ))
-        raise HTTPException(status_code=400, detail=f"Query blocked: {validation.blocked_reason}")
-
-    # Inject LIMIT
-    safe_sql = inject_limit(req.sql, req.row_limit)
-
-    # Check query cache (Feature #30) — same normalized query returns cached result
-    cached = query_cache.get(req.connection_name, req.sql, req.row_limit)
-    if cached:
-        await append_audit(AuditEntry(
-            id=str(uuid.uuid4()),
-            timestamp=time.time(),
-            event_type="query",
-            connection_name=req.connection_name,
-            sql=req.sql,
-            tables=cached.tables,
-            rows_returned=len(cached.rows),
-            duration_ms=0.0,
-            metadata={"cache_hit": True},
-        ))
-        return {
-            "rows": cached.rows,
-            "row_count": len(cached.rows),
-            "tables": cached.tables,
-            "execution_ms": cached.execution_ms,
-            "sql_executed": cached.sql_executed,
-            "cache_hit": True,
-        }
-
-    conn_str = get_connection_string(req.connection_name)
-    if not conn_str:
-        raise HTTPException(status_code=400, detail="No credentials stored for this connection")
-
-    # Acquire connector once for both cost estimation and query execution (perf optimization)
-    connector = await pool_manager.acquire(info.db_type, conn_str)
-
-    try:
-        # Cost estimation (Feature #13) — run EXPLAIN before execution
-        cost_estimate = None
-        try:
-            from .governance.cost_estimator import CostEstimator
-            cost_estimate = await CostEstimator.estimate(connector, safe_sql, info.db_type)
-
-            # Check budget before executing expensive queries
-            if cost_estimate.is_expensive and cost_estimate.estimated_usd > 0:
-                # Warn in audit log but don't block (policy-based blocking is a future feature)
-                await append_audit(AuditEntry(
-                    id=str(uuid.uuid4()),
-                    timestamp=time.time(),
-                    event_type="query",
-                    connection_name=req.connection_name,
-                    sql=req.sql,
-                    metadata={"cost_warning": True, "estimated_usd": cost_estimate.estimated_usd, "estimated_rows": cost_estimate.estimated_rows},
-                ))
-        except Exception:
-            pass  # Cost estimation is best-effort
-
-        start = time.monotonic()
-        try:
-            rows = await connector.execute(safe_sql, timeout=timeout)
-        except asyncio.TimeoutError:
-            health_monitor.record(req.connection_name, (time.monotonic() - start) * 1000, False, "timeout", info.db_type)
-            raise HTTPException(
-                status_code=408,
-                detail=f"Query timed out after {timeout}s. Consider adding more specific WHERE clauses or reducing the scope.",
-            )
-        except Exception as e:
-            health_monitor.record(req.connection_name, (time.monotonic() - start) * 1000, False, str(e)[:200], info.db_type)
-            sanitized = _sanitize_db_error(str(e))
-            raise HTTPException(status_code=500, detail=sanitized)
-    finally:
-        await pool_manager.release(info.db_type, conn_str)
-
-    elapsed_ms = (time.monotonic() - start) * 1000
-    health_monitor.record(req.connection_name, elapsed_ms, True, db_type=info.db_type)
-
-    # Apply PII redaction from annotations (Feature #15)
-    from .governance.pii import PIIRedactor
-    pii_redactor = PIIRedactor()
-    pii_columns = annotations.pii_columns
-    for col_name, rule in pii_columns.items():
-        pii_redactor.add_rule(col_name, rule)
-    if pii_redactor.has_rules():
-        rows = pii_redactor.redact_rows(rows)
-
-    # Store in cache after PII redaction (so cached data is already redacted)
-    query_cache.put(
-        connection_name=req.connection_name,
-        sql=req.sql,
-        row_limit=req.row_limit,
-        rows=rows,
-        tables=validation.tables,
-        execution_ms=elapsed_ms,
-        sql_executed=safe_sql,
-    )
-
-    # Charge query cost to budget (Feature #11 + #12)
-    query_cost_usd = (elapsed_ms / 1000) * 0.000014
-    budget_ledger.charge("default", query_cost_usd)
-
-    await append_audit(AuditEntry(
-        id=str(uuid.uuid4()),
-        timestamp=time.time(),
-        event_type="query",
-        connection_name=req.connection_name,
-        sql=req.sql,
-        tables=validation.tables,
-        rows_returned=len(rows),
-        duration_ms=elapsed_ms,
-        cost_usd=query_cost_usd,
-        metadata={"pii_redacted": pii_redactor.last_redacted_columns} if pii_redactor.last_redacted_columns else {},
-    ))
-
-    response = {
-        "rows": rows,
-        "row_count": len(rows),
-        "tables": validation.tables,
-        "execution_ms": elapsed_ms,
-        "sql_executed": safe_sql,
-        "cache_hit": False,
-        "pii_redacted": pii_redactor.last_redacted_columns if pii_redactor.last_redacted_columns else None,
-    }
-    if cost_estimate and not cost_estimate.warning:
-        response["cost_estimate"] = {
-            "estimated_rows": cost_estimate.estimated_rows,
-            "estimated_usd": round(cost_estimate.estimated_usd, 8),
-            "is_expensive": cost_estimate.is_expensive,
-        }
-    return response
-
-
-# ─── Audit ───────────────────────────────────────────────────────────────────
-
-@app.get("/api/audit")
-async def get_audit(
-    limit: int = Query(default=100, le=500),
-    offset: int = Query(default=0),
-    connection_name: str | None = None,
-    event_type: str | None = None,
-):
-    entries = await read_audit(
-        limit=limit,
-        offset=offset,
-        connection_name=connection_name,
-        event_type=event_type,
-    )
-    return {"entries": entries, "total": len(entries)}
-
-
-@app.get("/api/audit/export")
-async def export_audit(
-    connection_name: str | None = None,
-    event_type: str | None = None,
-    format: str = Query(default="json", pattern=r"^(json|csv)$"),
-):
-    """Export full audit trail for compliance (Feature #45).
-
-    Returns a downloadable JSON or CSV file with all audit entries
-    matching the filter criteria. Suitable for SOC 2, HIPAA, or EU AI Act reporting.
-    """
-    entries = await read_audit(
-        limit=10_000,
-        offset=0,
-        connection_name=connection_name,
-        event_type=event_type,
-    )
-
-    if format == "csv":
-        import csv
-        import io
-        output = io.StringIO()
-        writer = csv.writer(output)
-        writer.writerow([
-            "id", "timestamp", "event_type", "connection_name", "sql",
-            "tables", "rows_returned", "duration_ms", "blocked",
-            "block_reason", "agent_id", "metadata",
-        ])
-        for entry in entries:
-            e = entry if isinstance(entry, dict) else entry.__dict__
-            writer.writerow([
-                e.get("id", ""),
-                e.get("timestamp", ""),
-                e.get("event_type", ""),
-                e.get("connection_name", ""),
-                e.get("sql", ""),
-                ";".join(e.get("tables", [])),
-                e.get("rows_returned", ""),
-                e.get("duration_ms", ""),
-                e.get("blocked", False),
-                e.get("block_reason", ""),
-                e.get("agent_id", ""),
-                json.dumps(e.get("metadata", {})),
-            ])
-        content = output.getvalue()
-        return StreamingResponse(
-            iter([content]),
-            media_type="text/csv",
-            headers={"Content-Disposition": f"attachment; filename=signalpilot-audit-export.csv"},
-        )
-
-    # JSON format
-    export_data = {
-        "export_timestamp": time.time(),
-        "export_format": "signalpilot-audit-v1",
-        "filters": {
-            "connection_name": connection_name,
-            "event_type": event_type,
-        },
-        "entry_count": len(entries),
-        "entries": entries,
-    }
-    return StreamingResponse(
-        iter([json.dumps(export_data, indent=2, default=str)]),
-        media_type="application/json",
-        headers={"Content-Disposition": f"attachment; filename=signalpilot-audit-export.json"},
-    )
-
-
-# ─── Budget / Governance ────────────────────────────────────────────────────
-
-from .governance.budget import budget_ledger
 from .governance.cache import query_cache
 
-
-class BudgetCreateRequest(BaseModel):
-    session_id: str = Field(..., min_length=1, max_length=128)
-    budget_usd: float = Field(default=10.0, ge=0.01, le=10_000.0)
-
-
-@app.post("/api/budget", status_code=201)
-async def create_budget(req: BudgetCreateRequest):
-    """Create a budget for a session."""
-    budget = budget_ledger.create_session(req.session_id, req.budget_usd)
-    return budget.to_dict()
-
-
-@app.get("/api/budget/{session_id}")
-async def get_budget(session_id: str):
-    """Get budget status for a session."""
-    budget = budget_ledger.get_session(session_id)
-    if not budget:
-        raise HTTPException(status_code=404, detail="Session budget not found")
-    return budget.to_dict()
-
-
-@app.get("/api/budget")
-async def list_budgets():
-    """List all active session budgets."""
-    return {
-        "sessions": budget_ledger.get_all_sessions(),
-        "total_spent_usd": round(budget_ledger.total_spent, 6),
-    }
-
-
-@app.delete("/api/budget/{session_id}", status_code=204)
-async def close_budget(session_id: str):
-    """Close and remove a session budget."""
-    closed = budget_ledger.close_session(session_id)
-    if not closed:
-        raise HTTPException(status_code=404, detail="Session budget not found")
-
-
-# ─── Schema Annotations ────────────────────────────────────────────────────
-
-from .governance.annotations import load_annotations, generate_skeleton
-
-
-@app.get("/api/connections/{name}/annotations")
-async def get_annotations(name: str):
-    """Get schema annotations for a connection (Feature #16)."""
-    info = get_connection(name)
-    if not info:
-        raise HTTPException(status_code=404, detail=f"Connection '{name}' not found")
-    annotations = load_annotations(name)
-    return annotations.to_dict()
-
-
-@app.post("/api/connections/{name}/annotations/generate")
-async def generate_annotations(name: str):
-    """Generate a starter schema.yml from database introspection (Feature #29)."""
-    info = get_connection(name)
-    if not info:
-        raise HTTPException(status_code=404, detail=f"Connection '{name}' not found")
-
-    conn_str = get_connection_string(name)
-    if not conn_str:
-        raise HTTPException(status_code=400, detail="No credentials stored for this connection")
-
-    try:
-        connector = await pool_manager.acquire(info.db_type, conn_str)
-        try:
-            schema = await connector.get_schema()
-        finally:
-            await pool_manager.release(info.db_type, conn_str)
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=_sanitize_db_error(str(e)))
-
-    skeleton = generate_skeleton(schema, name)
-    return {
-        "connection_name": name,
-        "table_count": len(schema),
-        "yaml": skeleton,
-    }
-
-
-
-
-# ─── Query Cache ──────────────────────────────────────────────────────────────
-
-
-@app.get("/api/cache/stats")
-async def cache_stats():
-    """Get query cache statistics (Feature #30)."""
-    return query_cache.stats()
-
-
-@app.post("/api/cache/invalidate", status_code=200)
-async def invalidate_cache(connection_name: str | None = None):
-    """Invalidate cached query results. Optionally filter by connection."""
-    count = query_cache.invalidate(connection_name)
-    return {"invalidated": count, "connection_name": connection_name}
-
-
-@app.post("/api/connections/{name}/detect-pii")
-async def detect_pii(name: str):
-    """Auto-detect PII columns in a database schema based on naming patterns.
-
-    Returns suggested PII rules for columns with names matching known
-    PII patterns (email, ssn, phone, etc.). Results should be reviewed
-    and saved to schema.yml annotations.
-    """
-    info = get_connection(name)
-    if not info:
-        raise HTTPException(status_code=404, detail=f"Connection '{name}' not found")
-
-    conn_str = get_connection_string(name)
-    if not conn_str:
-        raise HTTPException(status_code=400, detail="No credentials stored for this connection")
-
-    # Get schema (from cache if available)
-    cached_schema = schema_cache.get(name)
-    if cached_schema is None:
-        try:
-            connector = await pool_manager.acquire(info.db_type, conn_str)
-            try:
-                cached_schema = await connector.get_schema()
-            finally:
-                await pool_manager.release(info.db_type, conn_str)
-            schema_cache.put(name, cached_schema)
-        except Exception as e:
-            raise HTTPException(status_code=500, detail=_sanitize_db_error(str(e)))
-
-    from .governance.pii import detect_pii_columns
-
-    all_detections: dict[str, dict[str, str]] = {}
-    for table_key, table_data in cached_schema.items():
-        columns = [col["name"] for col in table_data.get("columns", [])]
-        detected = detect_pii_columns(columns)
-        if detected:
-            all_detections[table_data.get("name", table_key)] = {
-                col: rule.value for col, rule in detected.items()
-            }
-
-    return {
-        "connection_name": name,
-        "tables_scanned": len(cached_schema),
-        "tables_with_pii": len(all_detections),
-        "detections": all_detections,
-    }
-
-
-@app.get("/api/schema-cache/stats")
-async def schema_cache_stats():
-    """Get schema cache statistics (Feature #18)."""
-    return schema_cache.stats()
-
-
-@app.post("/api/schema-cache/invalidate", status_code=200)
-async def invalidate_schema_cache(connection_name: str | None = None):
-    """Invalidate cached schema data. Optionally filter by connection."""
-    count = schema_cache.invalidate(connection_name)
-    return {"invalidated": count, "connection_name": connection_name}
-
-
-# ─── Metrics SSE ─────────────────────────────────────────────────────────────
 
 @app.get("/api/metrics")
 async def metrics_stream():
